third_party/blink/web_tests/external/wpt/fetch/api/basic/url-parsing.sub_encoding=windows-1252-expected.txt

This is a testharness.js-based test.
PASS Request uses the UTF-8 URL parser
PASS Request's referrer uses the UTF-8 URL parser
FAIL Response.redirect() uses the UTF-8 URL parser assert_equals: expected "%C3%BF" but got "%FF"
Harness: the test ran to completion.
